+++ bundles/mapping/mapwfs2/plugin/util/FeatureUtil.js
