+++ bundles/mapping/mapwfs2/plugin/util/defaultStyle.js
